+++ component/meta/MetaListFilter.js
@@ -13,6 +13,13 @@ module.exports = class MetaListFilter extends Base {
     }
 
     parse (data) {
+        if (data.items) {
+            return this.parseNestedItems(data);
+        }
+        if (this.class.key.name === data.attr) {
+            data.type = this.class.key.type;
+            return super.parse(data);
+        }
         const attr = this.class.getAttr(data.attr);
         if (!this.class.searchAttrs.includes(attr)) {
             return this.throwBadRequest(`Invalid search attribute: ${data.attr}.${this.class.id}`);
@@ -24,7 +31,31 @@ module.exports = class MetaListFilter extends Base {
             data.relation = true;
             data.valueType = attr.relation.getRefAttrType();
         }
+        if (attr.isClass()) {
+            data.type = 'class';
+        }
         return super.parse(data);
+    }
+
+    parseByType (data) {
+        switch (data.type) {
+            case 'class':
+                return this.parseClass(data);
+        }
+        return super.parseByType(data);
+    }
+
+    parseClass ({attr, op, value}) {
+        if (!value) {
+            return this.getEmptyValueCondition(attr, op);
+        }
+        const metaClass = this.class.meta.getClass(value);
+        if (!metaClass) {
+            return this.throwInvalidValue(value);
+        }
+        value = metaClass.isAbstract() ? [] : [metaClass.name];
+        value.push(...metaClass.getRealDescendants().map(({name}) => name));
+        return this.formatSelectorCondition(attr, op, value);
     }
 
     async parseRelation ({attr, op}) {