+++ component/other/Select2.js
@@ -69,7 +69,7 @@ module.exports = class Select2 extends Base {
     }
 
     getStringSearch (text) {
-        return new RegExp('^'+ EscapeHelper.escapeRegex(text), 'i');
+        return new RegExp(EscapeHelper.escapeRegex(text), 'i');
     }
 
     getItems (models) {